+++ src/Mapbender/CoreBundle/Resources/public/widgets/mapbender.popup.js
@@ -186,7 +186,22 @@
             $btn.addClass(confOrNode.cssClass || '');
             return $btn.get(0);
         },
-        __dummy__: null
+        /**
+         * make sure the popup never leaves the current viewport
+         */
+        restrictPositioning: function() {
+            this.$element.on('dragstop', function (event, ui) {
+                let forcedX = null;
+                let forcedY = null;
+                let $target = $(event.target);
+                if (ui.position.top < 0) forcedY = 0;
+                if (ui.position.top > window.innerHeight - 50) forcedY = window.innerHeight - 50;
+                if (ui.position.left < 0) forcedX = 0;
+                if (ui.position.left > window.innerWidth - 50) forcedX = window.innerWidth - 50;
+                if (forcedX !== null) $target.css('left', forcedX);
+                if (forcedY !== null) $target.css('top', forcedY);
+            });
+        },
     };
 
     window.Mapbender.Popup = function Popup(options) {
@@ -333,6 +348,7 @@
                     containment: containment,
                     scroll: false
                 });
+                this.restrictPositioning();
             }
             this.focus();
         },
@@ -554,6 +570,7 @@
                 if (draggableOptions) {
                     self.$element.css('position', 'relative');
                     self.$element.draggable(draggableOptions);
+                    self.restrictPositioning();
                 }
                 self.$element.trigger('openend');
             }, 100);